Revert "Implement event-driven popup management; refactor sound settings and …"
