--- assets/Script/Controller.meta
@@ -1,7 +1,0 @@
-{
-  "ver": "1.0.1",
-  "uuid": "9ee770bd-44d3-4ff8-93aa-0fd8ef3c3aa7",
-  "isSubpackage": false,
-  "subpackageName": "",
-  "subMetas": {}
-}
