--- assets/Script/Controller/LobbyController.js
@@ -1,20 +1,0 @@
-const EventBus = require('Core/EventBus');
-const PopupType = require('Enum/PopupType');
-
-let LobbyController = cc.Class({
-    extends: cc.Component,
-
-    onLoad() {
-        LobbyController.instance = this;
-    },
-
-    onClickSetting() {
-        EventBus.emit('SHOW_POPUP', PopupType.SETTING);
-    },
-
-    onClickRanking() {
-        EventBus.emit('SHOW_POPUP', PopupType.RANKING);
-    }
-});
-
-module.exports = LobbyController;
