--- assets/Script/Core.meta
@@ -1,7 +1,0 @@
-{
-  "ver": "1.0.1",
-  "uuid": "dc51f2cb-bd9d-4817-9fe5-6464408a1896",
-  "isSubpackage": false,
-  "subpackageName": "",
-  "subMetas": {}
-}
