--- assets/Script/Enum.meta
@@ -1,7 +1,0 @@
-{
-  "ver": "1.0.1",
-  "uuid": "747e4f04-ceb2-4fce-b191-f00125ece584",
-  "isSubpackage": false,
-  "subpackageName": "",
-  "subMetas": {}
-}
