--- assets/Script/Enum/PopupType.js
@@ -1,6 +1,0 @@
-const PopupType = {
-    SETTING: 'SETTING',
-    RANKING: 'RANKING'
-};
-Object.freeze(PopupType);
-module.exports = PopupType;
