--- assets/Script/Enum/PopupType.js.meta
@@ -1,9 +1,0 @@
-{
-  "ver": "1.0.5",
-  "uuid": "a4b854fe-dfab-497c-a80c-ef04485ec891",
-  "isPlugin": false,
-  "loadPluginInWeb": true,
-  "loadPluginInNative": true,
-  "loadPluginInEditor": false,
-  "subMetas": {}
-}
