+++ assets/Script/Pupop/RankingPopup.js
@@ -4,7 +4,7 @@ cc.Class({
     properties: {
         rankingTable: {
             default: null,
-            type: require('RankingTableController')
+            type: require('RankingTable'),
         },
         data: {
             default: [],
